requirements.txt
python-dotenv
mysql
mysql-connector-python
stringcase